[libromdata] SegaPVR: Cast sz_header to uint32_t before using it in DetectInfo.
This fixes an MSVC warning:
SegaPVR.cpp(636): warning C4267: '=': conversion from 'size_t' to 'uint32_t', possible loss of data

+++ src/libromdata/Texture/SegaPVR.cpp
@@ -633,7 +633,7 @@ SegaPVR::SegaPVR(IRpFile *file)
 	// Check if this PVR image is supported.
 	DetectInfo info;
 	info.header.addr = 0;
-	info.header.size = sz_header;
+	info.header.size = (uint32_t)sz_header;
 	info.header.pData = header;
 	info.ext = nullptr;	// Not needed for PVR.
 	info.szFile = file->size();
